Move getkey.py to tools folder
