+++ bushexa/chroniccrawler/crawler/buspos.py
@@ -1,7 +1,7 @@
 import requests
 import xmltodict
 
-from .getkey import get_key
+from .tools.getkey import get_key
 
 from chroniccrawler.models import LaneToTrack, PosOfBus
