+++ bushexa/chroniccrawler/crawler/dayinfo.py
@@ -3,7 +3,7 @@ import xmltodict
 
 import datetime
 
-from .getkey import get_key
+from .tools.getkey import get_key
 
 from chroniccrawler.models import DayInfo
