+++ bushexa/chroniccrawler/crawler/laneinfo.py
@@ -2,7 +2,7 @@ import requests
 import itertools
 import xmltodict
 
-from .getkey import get_key
+from .tools.getkey import get_key
 
 from chroniccrawler.models import LaneToTrack, NodeOfLane
