+++ bushexa/chroniccrawler/crawler/timetable_usb.py
@@ -1,7 +1,7 @@
 import requests
 import xmltodict
 
-from .getkey import get_key
+from .tools.getkey import get_key
 
 from chroniccrawler.models import UlsanBus_LaneToTrack, UlsanBus_TimeTable
